front: fix timezone
If the user is not on the same timezone as the server, the date getting from server take the hour diff in account.
Problem: 01/11 00h from the server become 31/10 23h00 after using new Date(). So if we print the current month

To avoid that, we have to fix the timezone on the client side.